Add files via upload
The Sales Data given belongs to Electronic Stores across 10 cities of The United States of America. The following project depicts the business insights which is very easy to understand. It consists of city wise and month wise revenue generated, total number of orders, average sales, the top selling products, city wise sales, month wise sales, and percentage contribution by each of the cities. Using these insights, one can focus on increasing the profitability with the top products and the necessary changes needed in order to boost the low revenue generation by products/city/month.
So far, San Francisco and Los Angeles are the cities making the highest sales, with the top products being Macbook, Thinkpad and iPhone. The total revenue generated is 63.76000 US Dollars throughout the year. The highest sales is made in the month of August and November being the least sales generating month.