Add files via upload
The Diabetes data is studied and is further used for building predictive models which in turn can be used in the medical fields in order to know if a patient is suspected to have diabetes or not. 
There are a lot of interesting insights found here:
1) The chances of being diabetic increases if the number of pregnancies are increased
2) It is seen that at younger ages between 20-30 years, if the glucose level is in between 75-150, it is okay, the patient is safe from diabetes. But if the same glucose is maintained at older ages, there might be chances of diabetes. Hence, as the age increases, the glucose should be reduced in order to avoid diabetes
3) As the skin thickness increases, the probabilty of being a diabeteic patient increases
4) As the age increases, the skin thickness decreases
5) The BP is usually increasing with the age, but there is a sudden dip in blood pressure after 70 years of age
6) In younger ages, if the diabetes pedigree function is low, there are low chances of diabetes. But as the age advances, having a lower diabetes pedigree function also doesn't help in getting exposed to diabetes. This means that this function's value changes according to age

In the predictive model building, out of the supervised machine learning algorithms, Random Forest Classifier gives us the best accuracy of 82% with a fair amount of efforts to predict diabetes positive and diabetes negative in various cases.