Add files via upload
Insights of Attrition in a company from the HR Department:
1) 16.1% percent of the employees tend to quit the company whereas 83.9% stick on to the same
2) It is seen that most of the people who quit the job are single
Whereas the people who stick on the most are married. This might be to avoid new location to work or a new location to relocate or the facilities given by the company are beneficial than other companies here.
3) It is seen that Laboratory technicians have more attrition and Sales executives, most people do not tend to quit.
This might be because of variation in staff incentives, or treatment of employees
4) Male attrition rate is higher than female attrition rate
5) More attrition is seen by people who travel very rarely for the company business
6) The above graph tells us that most people with an experience of 1 year tend to change the company
7) Attrition is higher in 3rd grade education
8) The attrition is more where experience was 1 year. So pretty obviously, the number of jobs changed within 1 year will also be 1 which is also having a higher attrition numbers.
9) People who stay closer to the work place tend to change/shift their jobs more
If these parameters are taken care of while hiring, then the number of people quitting and if the treatment or incentives to laboratory technicians, sales executives and research scientists are improved, the attrition can be reduced.